+++ lib/screens/cases/case_info/case_history_page.dart
@@ -189,6 +189,7 @@ class _CaseHistoryPageState extends State<CaseHistoryPage>
             initialStage: proceeding.nextStageId,
             stageList: stageList,
             insertedBy: proceeding.insertedBy,
+            initialRemark: proceeding.remark,
           );
         },
       );